# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for ليلتك wedding venue platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /bookings:
    get:
      operationId: listBookings
      tags: [bookings]
      summary: List bookings for the current user
      parameters:
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
    post:
      operationId: createBooking
      tags: [bookings]
      summary: Create a booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingInput"
      responses:
        "201":
          description: Booking created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "400":
          description: Bad request
  /bookings/{id}:
    get:
      operationId: getBooking
      tags: [bookings]
      summary: Get booking details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "404":
          description: Not found
    patch:
      operationId: updateBooking
      tags: [bookings]
      summary: Update a booking (e.g. cancel)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingUpdate"
      responses:
        "200":
          description: Updated booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
...
  /invitations:
    post:
      operationId: createInvitation
      tags: [invitations]
      summary: Create a new e-invitation (returns public + manage tokens)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvitationInput"
      responses:
        "201":
          description: Invitation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invitation"
  /invitations/public/{publicToken}:
    get:
      operationId: getPublicInvitation
      tags: [invitations]
      summary: Get the public invitation view for a guest (no manage data)
      parameters:
        - name: publicToken
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Public invitation details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicInvitation"
        "404":
          description: Not found
  /invitations/public/{publicToken}/rsvp:
    post:
      operationId: respondToInvitation
      tags: [invitations]
      summary: Submit a guest RSVP (attending or not, with party size)
      parameters:
        - name: publicToken
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RsvpInput"
      responses:
        "201":
          description: RSVP recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rsvp"
        "404":
          description: Not found
  /invitations/manage:
    get:
      operationId: getManagedInvitation
      tags: [invitations]
      summary: Get the host dashboard view (invitation + RSVPs + stats)
      description: >
        Authenticates via `Authorization: Bearer <manageToken>`.
        The token must never appear in the URL path or query string.
      security:
        - ManageBearerAuth: []
      responses:
        "200":
          description: Managed invitation with RSVP list and stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManagedInvitation"
        "401":
          description: Missing or invalid bearer token
        "404":
          description: Not found

  # ─── REVIEWS ─────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Booking:
      type: object
      required: [id, venueId, venueName, eventDate, guestCount, status, totalPrice, reviewToken, createdAt]
      properties:
        id:
          type: integer
        venueId:
          type: integer
        venueName:
          type: string
        venueImage:
          type: ["string", "null"]
        eventDate:
          type: string
        guestCount:
          type: integer
        status:
          type: string
          enum: [pending, confirmed, cancelled, completed]
        totalPrice:
          type: number
        reviewToken:
          type: string
          description: Unguessable UUID token used to submit a post-stay review (one-time use)
        notes:
          type: ["string", "null"]
        groomName:
          type: ["string", "null"]
        brideName:
          type: ["string", "null"]
        contactPhone:
          type: ["string", "null"]
        createdAt:
          type: string
    BookingInput:
      type: object
      required: [venueId, eventDate, guestCount, groomName, brideName, contactPhone]
      properties:
        venueId:
          type: integer
        eventDate:
          type: string
          pattern: '^\d{4}-\d{2}-\d{2}$'
        guestCount:
          type: integer
          minimum: 1
          maximum: 10000
        notes:
          type: string
          maxLength: 1000
        groomName:
          type: string
          minLength: 1
          maxLength: 200
        brideName:
          type: string
          minLength: 1
          maxLength: 200
        contactPhone:
          type: string
          minLength: 5
          maxLength: 20
    BookingUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [pending, confirmed, cancelled, completed]
        notes:
          type: string
          maxLength: 1000
    InvitationInput:
      type: object
      required: [inviterType, eventDate]
      properties:
        template:
          type: string
          enum: [emerald, classic]
        inviterType:
          type: string
          enum: [groom, bride_family]
        groomName:
          type: ["string", "null"]
          maxLength: 100
        groomFatherName:
          type: ["string", "null"]
          maxLength: 100
        brideFatherName:
          type: ["string", "null"]
          maxLength: 100
        brideMotherName:
          type: ["string", "null"]
          maxLength: 100
        groomSideRef:
          type: ["string", "null"]
          maxLength: 100
        eventDate:
          type: string
          maxLength: 20
        eventTime:
          type: ["string", "null"]
          maxLength: 50
        venueName:
          type: ["string", "null"]
          maxLength: 200
        venueAddress:
          type: ["string", "null"]
          maxLength: 500
        note:
          type: ["string", "null"]
          maxLength: 1000
        giftEnabled:
          type: boolean
        giftIban:
          type: ["string", "null"]
          maxLength: 50
        giftBankName:
          type: ["string", "null"]
          maxLength: 100
        giftStcPay:
          type: ["string", "null"]
          maxLength: 50
        giftNote:
          type: ["string", "null"]
          maxLength: 500
    Invitation:
      type: object
      required: [id, publicToken, manageToken, inviterType, eventDate, createdAt]
      properties:
        id:
          type: integer
        publicToken:
          type: string
        manageToken:
          type: string
        template:
          type: string
          enum: [emerald, classic]
        inviterType:
          type: string
          enum: [groom, bride_family]
        groomName:
          type: ["string", "null"]
        groomFatherName:
          type: ["string", "null"]
        brideFatherName:
          type: ["string", "null"]
        brideMotherName:
          type: ["string", "null"]
        groomSideRef:
          type: ["string", "null"]
        eventDate:
          type: string
        eventTime:
          type: ["string", "null"]
        venueName:
          type: ["string", "null"]
        venueAddress:
          type: ["string", "null"]
        note:
          type: ["string", "null"]
        giftEnabled:
          type: boolean
        giftIban:
          type: ["string", "null"]
        giftBankName:
          type: ["string", "null"]
        giftStcPay:
          type: ["string", "null"]
        giftNote:
          type: ["string", "null"]
        createdAt:
          type: string
    PublicInvitation:
      type: object
      required: [publicToken, inviterType, eventDate]
      properties:
        publicToken:
          type: string
        template:
          type: string
          enum: [emerald, classic]
        inviterType:
          type: string
          enum: [groom, bride_family]
        groomName:
          type: ["string", "null"]
        groomFatherName:
          type: ["string", "null"]
        brideFatherName:
          type: ["string", "null"]
        brideMotherName:
          type: ["string", "null"]
        groomSideRef:
          type: ["string", "null"]
        eventDate:
          type: string
        eventTime:
          type: ["string", "null"]
        venueName:
          type: ["string", "null"]
        venueAddress:
          type: ["string", "null"]
        note:
          type: ["string", "null"]
        giftEnabled:
          type: boolean
        giftIban:
          type: ["string", "null"]
        giftBankName:
          type: ["string", "null"]
        giftStcPay:
          type: ["string", "null"]
        giftNote:
          type: ["string", "null"]
    RsvpInput:
      type: object
      required: [guestName, attending]
      properties:
        guestName:
          type: string
          minLength: 1
          maxLength: 100
        attending:
          type: boolean
        partySize:
          type: integer
          minimum: 1
          maximum: 20
        message:
          type: ["string", "null"]
          maxLength: 500
    Rsvp:
      type: object
      required: [id, guestName, attending, partySize, createdAt]
      properties:
        id:
          type: integer
        guestName:
          type: string
        attending:
          type: boolean
        partySize:
          type: integer
        message:
          type: ["string", "null"]
        createdAt:
          type: string
    ManagedInvitation:
      type: object
      required: [invitation, rsvps, stats]
      properties:
        invitation:
          $ref: "#/components/schemas/Invitation"
        rsvps:
          type: array
          items:
            $ref: "#/components/schemas/Rsvp"
        stats:
          type: object
          required: [totalResponses, attendingResponses, decliningResponses, totalAttendingGuests]
          properties:
            totalResponses:
              type: integer
            attendingResponses:
              type: integer
            decliningResponses:
              type: integer
            totalAttendingGuests:
              type: integer
...
  securitySchemes:
    ManageBearerAuth:
      type: http
      scheme: bearer
      description: >
        Invitation management bearer token (manageToken).
        Send as `Authorization: Bearer <token>`. Never place this token in
        a URL path or query string — it must only travel as a request header.