# lib/api-zod/src/generated/types/bookingUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for ليلتك wedding venue platform
 * OpenAPI spec version: 0.1.0
 */
import type { BookingUpdateStatus } from './bookingUpdateStatus';

export interface BookingUpdate {
  status?: BookingUpdateStatus;
  /** @maxLength 1000 */
  notes?: string;
}